Open "application/vnd.google-apps.form" as Spreadsheet instead of generic File.
Forms powered by Google Spreadsheets are accessible over the Spreadsheets API. We have a few API clients that rely on this behavior (aggregating data in some of our hand-edited spreadsheets along with some data in user-facing Google Forms). I *think* this worked in some of our older code with `google-spreadsheet-ruby` (maybe it wasn't checking MIME, or Google hadn't changed the MIME at that point).

Anwyay, this allows us to access a Form by `spreadsheet_by_key` without getting

```GoogleDrive::Error - The file with the ID is not a spreadsheet: ```

…and also allows us to use the rest of the Spreadsheet API to retrieve the data in that resultant spreadsheet.